refactor(stats): flip FE to /occurrences/stats/, delete legacy backend
Completes the migration started in Commit 2:
- Move the frontend hook to ui/src/data-services/hooks/occurrences/stats/
  (mirrors the backend URL hierarchy)
- Flip the hook URL to /occurrences/stats/top-identifiers/
- Delete UserIdentificationCountsView class + its path() registration
  at /users/identifications/top/

Response envelope is unchanged, so no UI logic changes — just the URL.
